requirements.txt
# requirements.txt
google-generativeai>=0.3.0
mysql-connector-python>=8.0.0
langchain>=0.1.0
langchain-community>=0.0.1
Flask==3.0.0
Flask-Session==0.5.0
mysql-connector-python==8.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1